JSON Serialization
Serialization for class Person's attributes with JSON
