--- RestWithSpringboot/comandos para flyway.txt
@@ -1,10 +1,0 @@
-mvn clean package spring-boot:run
-
-// realiza o clean package do maven e gera o jar, em seguida executa o spring-boot para executar as migrations
-
-
-----------------------------------------------------------------------------
-
-mvn flyway:migrate
-
-//para executar flyway sem precisar executar springboot
