+++ RestWithSpringboot/src/main/java/br/com/wellmmjr/controller/PersonController.java
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.com.wellmmjr.data.vo.PersonVO;
+import br.com.wellmmjr.data.vo.v1.PersonVO;
 import br.com.wellmmjr.services.PersonServices;
 
 @RestController
-@RequestMapping("/person")
+@RequestMapping("/api/person/v1")
 public class PersonController {
 	
 	@Autowired
