--- RestWithSpringboot/src/main/java/br/com/wellmmjr/data/vo/PersonVO.java
@@ -1,75 +1,0 @@
-package br.com.wellmmjr.data.vo;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-public class PersonVO implements Serializable {
-
-	private static final long serialVersionUID = 1L;
-
-	private Long id;
-	
-	private String firstName;
-	
-	private String secondName;
-	
-	private String address;
-	
-	private String gender;
-	
-	public PersonVO() {
-		
-	}
-	
-	public Long getId() {
-		return id;
-	}
-	public void setId(Long id) {
-		this.id = id;
-	}
-	public String getFirstName() {
-		return firstName;
-	}
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-	public String getSecondName() {
-		return secondName;
-	}
-	public void setSecondName(String secondName) {
-		this.secondName = secondName;
-	}
-	public String getAddress() {
-		return address;
-	}
-	public void setAddress(String address) {
-		this.address = address;
-	}
-	public String getGender() {
-		return gender;
-	}
-	public void setGender(String gender) {
-		this.gender = gender;
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(address, firstName, gender, id, secondName);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		PersonVO other = (PersonVO) obj;
-		return Objects.equals(address, other.address) && Objects.equals(firstName, other.firstName)
-				&& Objects.equals(gender, other.gender) && id == other.id
-				&& Objects.equals(secondName, other.secondName);
-	}
-
-	
-}
