+++ RestWithSpringboot/src/main/java/br/com/wellmmjr/data/vo/v1/PersonVO.java
@@ -1,0 +1,80 @@
+package br.com.wellmmjr.data.vo.v1;
+
+import java.util.Objects;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonPropertyOrder({"address", "gender", "first_name", "last_name", "id"})
+public class PersonVO {
+
+	private long id;
+	
+	@JsonProperty("first_name")
+	private String firstName;
+	
+	@JsonProperty("last_name")
+	private String secondName;
+	
+	@JsonIgnore
+	private String address;
+	
+	private String gender;
+	
+	public PersonVO() {
+		
+	}
+	
+	public long getId() {
+		return id;
+	}
+	public void setId(long id) {
+		this.id = id;
+	}
+	public String getFirstName() {
+		return firstName;
+	}
+	public void setFirstName(String firstName) {
+		this.firstName = firstName;
+	}
+	public String getSecondName() {
+		return secondName;
+	}
+	public void setSecondName(String secondName) {
+		this.secondName = secondName;
+	}
+	public String getAddress() {
+		return address;
+	}
+	public void setAddress(String address) {
+		this.address = address;
+	}
+	public String getGender() {
+		return gender;
+	}
+	public void setGender(String gender) {
+		this.gender = gender;
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(address, firstName, gender, id, secondName);
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		PersonVO other = (PersonVO) obj;
+		return Objects.equals(address, other.address) && Objects.equals(firstName, other.firstName)
+				&& Objects.equals(gender, other.gender) && id == other.id
+				&& Objects.equals(secondName, other.secondName);
+	}
+
+	
+}
