+++ RestWithSpringboot/src/main/resources/application.properties
@@ -6,9 +6,3 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialec
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=false
-
-#Flyway Configs
-
-flyway.url=jdbc:mysql://localhost:3306/
-flyway.schemas=rest_with_springboot
-flyway.user=root
