--- RestWithSpringboot/src/main/resources/db/migration/V2__Populate_Table_Person.sql
@@ -1,4 +1,0 @@
-INSERT INTO `person` (`id`, `address`, `first_name`, `gender`, `second_name`) VALUES
-  (2, 'just a rainbow in the dark', 'Wellington', 'non-binary', 'M. Matias Junior'),
-  (4, 'far away from me so far i just cant see', 'Milena', 'non-binary', 'Cravo e Canela'),
-  (5, 'baixada', 'Cris', 'non-binary', 'Muller');
